Update Poin to Double, Rev Pop Up Verification, Article on Top Slider

See merge request joe43/jlf-mobile!37

+++ lib/models/point_history.dart
@@ -11,7 +11,7 @@ String pointHistoryToJson(List<PointHistory> data) => json.encode(List<dynamic>.
 class PointHistory {
     int id;
     String information;
-    int point;
+    double point;
     int userId;
     dynamic animalId;
     dynamic type;
